requirements.txt
beautifulsoup4==4.8.0
dj-database-url==0.5.0
Django==2.2.5
django-bootstrap4==1.0.1
django-fontawesome-5==1.0.16
django-heroku==0.3.1
django-phone-field==1.8.0
psycopg2==2.8.3
pytz==2019.2
PyYAML==5.1.2
soupsieve==1.9.3
sqlparse==0.3.0
whitenoise==4.1.3
